추가수정

+++ Assets/2DPlatformerRND/01_Scripts/Managers/ItemStats.cs
@@ -5,16 +5,16 @@ namespace PahlBit
     public class ItemStats
     {
         public Percent HealthUp;
-        public double HealthRegen;
+        public float HealthRegen;
         public Percent ManaUp;
-        public double ManaRegen;
+        public float ManaRegen;
         public Percent AttackUp;
         public Percent DefenceUp;
         public Percent MoveSpeedUp;
         public Percent AttackSpeedUp;
         public Percent CooltimeDown;
-        public double ShieldAdd;
-        public double ShieldRegen;
+        public float ShieldAdd;
+        public float ShieldRegen;
         public Percent FireResist;
         public Percent IceResist;
         public Percent LightningResist;
